// src/constants/enums.ts
export enum Genres {
  ALL = "All",
  COMEDY = "Comedy",
  DRAMA = "Drama",
  ACTION = "Action",
  ROMANCE = "Romance",
}

export enum PARAMS {
  LIMIT = "limit",
  SEARCH_BY = "searchBy",
  SORT_ORDER = "sortOrder",
  SORT_BY = "sortBy",
  SEARCH = "search",
  FILTER = "filter",
}

export enum DEFAULT_QUERIES {
  RELEASE_DATE = "release_date",
  TITLE = "title",
}